aruha-729: almost done. refactored instance creation and moved everything in one file
