+++ tests/instance_control/ebs_config.json
@@ -1,26 +1,26 @@
 [
   {
+    "account": "aruha-test",
     "cluster_name": "bubuku-staging-esb-mount-15",
     "regions": ["eu-central-1"],
     "vpc_id": "vpc-0d7c4564",
-    "odd_sg_id": "sg-8c47ebe4",
     "cluster_size": 1,
     "instance_type": "m4.large",
     "volume_type": "st1",
     "volume_size": 500,
     "scalyr_key": "0xRvOECatMkiGFqpSV2YFQMuDO6Hi6aRDPUQ7gUglXFY-",
     "image_version": "1.2.2.2-kafka-10-jdk",
-    "environment": [
-      "ZOOKEEPER_STACK_NAME=exhibitor-staging",
-      "HEALTH_PORT=8888",
-      "BUBUKU_MODE=amazon",
-      "zookeeper_prefix=/stagingebsmount",
-      "buku_features=restart_on_exhibitor,graceful_terminate,use_ip_address",
-      "jmx_port=8004",
-      "kafka_heap_opts=-Xmx6G -Xms6G",
-      "startup_timeout_type=linear",
-      "startup_timeout_initial=300",
-      "startup_timeout_step=60"
-    ]
+    "environment": {
+      "ZOOKEEPER_STACK_NAME": "exhibitor-staging",
+      "HEALTH_PORT": "8888",
+      "BUBUKU_MODE": "amazon",
+      "zookeeper_prefix": "/staging",
+      "buku_features": "restart_on_exhibitor,graceful_terminate,use_ip_address",
+      "jmx_port": 8004,
+      "kafka_heap_opts": "-Xmx6G -Xms6G",
+      "startup_timeout_type": "linear",
+      "startup_timeout_initial": 300,
+      "startup_timeout_step": 60
+    }
   }
 ]
